API Gateway creation and combination

+++ stall-management-service/src/main/java/com/bookfair/stallmanagement/dto/StallResponse.java
@@ -7,6 +7,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
@@ -38,7 +39,7 @@ public class StallResponse {
     
     private Double length;
     
-    private Double pricePerDay;
+    private BigDecimal pricePerDay;
     
     private String description;
